Move `sha256sum` into `project_helpers.py`
